librats: fix compile-time printf warning

+++ attesters/csv/collect_evidence.c
@@ -221,7 +221,7 @@ static int collect_attestation_evidence(const uint8_t *hash, uint32_t hash_len,
 	user_data_pa = (uint64_t)gva_to_gpa(user_data);
 	ret = do_hypercall(KVM_HC_VM_ATTESTATION, (unsigned long)user_data_pa, CSV_GUEST_MAP_LEN);
 	if (ret) {
-		RATS_ERR("failed to save attestation report to %#016lx (ret:%d)\n", ret,
+		RATS_ERR("failed to save attestation report to %#016lx (ret:%lu)\n", (unsigned long)ret,
 			 user_data_pa);
 		goto err_munmap;
 	}
@@ -275,7 +275,7 @@ rats_attester_err_t csv_collect_evidence(rats_attester_ctx_t *ctx, attestation_e
 
 	snprintf(evidence->type, sizeof(evidence->type), "csv");
 
-	RATS_DEBUG("ctx %p, evidence %p, report_len %d\n", ctx, evidence, evidence->csv.report_len);
+	RATS_DEBUG("ctx %p, evidence %p, report_len %u\n", ctx, evidence, evidence->csv.report_len);
 
 	return RATS_ATTESTER_ERR_NONE;
 }
